feat: navbar styling
