+++ frontend/src/app/globals.css
@@ -11,7 +11,7 @@
   --color-background-cream:         #FEF4C7; /* Main page background */
   --color-background-cream-rgb:     254, 244, 199;
 
-  --color-surface-light-cream:      #FEF9E7; /* Navbar, lighter sections */
+  --color-surface-light-cream:      #FEF9E7; /* NavBar, lighter sections */
   --color-surface-light-cream-rgb:  254, 249, 231;
 
   --color-surface-extra-light:      #FFFBEB; /* Hover states, card backgrounds */
